+++ apps/node/tests/test11.ts
@@ -1,6 +1,11 @@
 import { Assets } from '../index.js';
-import { PDFDocument, PDFFont, StandardFonts } from '../../../dist/es/index.js';
-import { charAtIndex, last } from '../../../dist/es/utils/index.js';
+import {
+  charAtIndex,
+  last,
+  PDFDocument,
+  PDFFont,
+  StandardFonts,
+} from '../../../dist/es/all.js';
 
 const breakTextIntoLines = (
   text: string,
