+++ apps/node/tests/test14.ts
@@ -1,7 +1,6 @@
 import { Assets } from '../index.js';
-import { PDFDocument, PDFTextField } from '../../../dist/es/index.js';
+import { PDFDocument, PDFTextField, values } from '../../../dist/es/all.js';
 import type { PDFField } from '../../../dist/es/api/form/PDFField.js';
-import { values } from '../../../dist/es/utils/index.js';
 
 const fieldNames = {
   // Page 1
