+++ apps/node/tests/test2.ts
@@ -4,7 +4,7 @@ import {
   PDFDocument,
   PDFPage,
   rgb,
-} from '../../../dist/es/index.js';
+} from '../../../dist/es/all.js';
 
 // This test loads an existing PDF document with many pages.
 // It inserts data for every page (images, rectangles, texts, embedded PDFs).
