+++ apps/node/tests/test20.ts
@@ -1,5 +1,5 @@
 import { Assets } from '../index.js';
-import { PDFDocument } from '../../../dist/es/index.js';
+import { PDFDocument } from '../../../dist/es/all.js';
 
 // Based on test19.ts
 export default async (assets: Assets) => {
