+++ apps/node/tests/test4.ts
@@ -7,7 +7,7 @@ import {
   StandardFonts,
   rgb,
   degrees,
-} from '../../../dist/es/index.js';
+} from '../../../dist/es/all.js';
 
 export default async (assets: Assets) => {
   const { pdfs, images } = assets;
