+++ apps/node/tests/test6.ts
@@ -4,7 +4,7 @@ import {
   ParseSpeeds,
   PDFDocument,
   rgb,
-} from '../../../dist/es/index.js';
+} from '../../../dist/es/all.js';
 
 export default async (assets: Assets) => {
   const { pdfs, images, fonts } = assets;
