+++ apps/node/tests/test8.ts
@@ -5,7 +5,7 @@ import {
   PDFPage,
   rgb,
   StandardFonts,
-} from '../../../dist/es/index.js';
+} from '../../../dist/es/all.js';
 
 export default async (assets: Assets) => {
   const { pdfs } = assets;
